Preprocessing Jupyter Notebooks & Output Files
Here are the Jupyter Notebooks used to preprocess our data.  There is one notebook for each dataset we used that runs the preprocessing for the N-gram model and a single notebook that runs all of the preprocessing for the character based LSTM model.  Also uploaded are the .out and .txt files outputted by the notebook that contain the preprocessed data.